Make sample_rate f32

+++ src/octaver/octaver.rs
@@ -12,7 +12,7 @@ pub struct Octaver {
 }
 
 impl Octaver {
-  pub fn new(sample_rate: f64) -> Self {
+  pub fn new(sample_rate: f32) -> Self {
     Self {
       lowpass: Lowpass::new(sample_rate),
       delta: Delta::new(),
